project compliance fixes
